Split branch manager into seperate traits
