+++ crates/gitbutler-branch-actions/src/actions.rs
@@ -17,12 +17,13 @@ use gitbutler_reference::{Refname, RemoteRefname};
 use gitbutler_repo::{credentials::Helper, RepoActions, RepositoryExt};
 use std::path::Path;
 
+use crate::branch_manager::branch_removal::BranchRemoval;
 use crate::{
     base::{
         get_base_branch_data, set_base_branch, set_target_push_remote, update_base_branch,
         BaseBranch,
     },
-    branch_manager::BranchManagerAccess,
+    branch_manager::{branch_creation::BranchCreation, BranchManagerAccess},
     remote::{get_branch_data, list_remote_branches, RemoteBranch, RemoteBranchData},
     VirtualBranchesExt,
 };
